Using an abstract class for render passes now.
