+++ game-engine/src/main/java/gameengine/engine/renderer/ScenePass.java
@@ -1,8 +1,5 @@
 package gameengine.engine.renderer;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.lwjgl.opengl.GL46;
 
 import gameengine.engine.IGameObject;
@@ -11,22 +8,15 @@ import gameengine.engine.renderer.shader.ShaderProgram;
 import gameengine.engine.renderer.uniform.UniformManager;
 import gameengine.util.FileUtils;
 
-public class ScenePass implements IRenderPass<IGameObject> {
+public class ScenePass extends ARenderPass<IGameObject> {
 	public static final String VERTEX_SHADER = "shd-scene-vert";
 	public static final String FRAGMENT_SHADER = "shd-scene-frag";
 	
 	Projection projection;
     Camera camera;
-	
-	private List<IGameObject> submissions;
-    private ShaderProgram shaderProgram;
-    private UniformManager uniformsMap;
 
     public ScenePass() {
-    	this.submissions = new ArrayList<>();
-    	this.shaderProgram = null;
-    	this.projection = null;
-    	this.camera = null;
+    	super();
     }
     
     
@@ -49,15 +39,9 @@ public class ScenePass implements IRenderPass<IGameObject> {
     	this.shaderProgram.generate();
     	this.createUniforms();
     }
-    
-    @Override
-    public void reset() {
-    	this.submissions.clear();
-    }
 
     @Override
     public void execute() {
-    	System.out.println(this.submissions.size());
         this.shaderProgram.bind();
         this.uniformsMap.setUniform("txtSampler", 0);
         this.uniformsMap.setUniform("projectionMatrix", this.projection.getProjMatrix());
@@ -75,15 +59,4 @@ public class ScenePass implements IRenderPass<IGameObject> {
     public void dispose() {
         this.shaderProgram.dispose();
     }
-    
-    @Override
-    public void submit(IGameObject object) {
-    	this.submissions.add(object);
-    }
-
-
-	@Override
-	public UniformManager getUniformManager() {
-		return this.uniformsMap;
-	}
 }
